move theme.ts out of pages
