+++ packages/website/src/pages/token.tsx
@@ -1,17 +1,5 @@
-import Head from "next/head";
-import {
-  Box,
-  Flex,
-  Button,
-  Heading,
-  Text,
-  Link,
-  Image,
-  List,
-  Stack,
-} from "@chakra-ui/react";
+import { Box } from "@chakra-ui/react";
 import Layout from "../components/layout/Layout";
-import { Footer } from "../components/layout/Footer";
 import { TokenSection } from "../components/Section/TokenSection";
 
 export default function token() {
